requirements.txt
tensorflow-gpu==2.3.0
opencv-python
numpy
tqdm
pillow